requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
pypdf==3.17.1
python-docx==1.1.0
openai==1.3.7
google-generativeai==0.3.2 